requirements.txt
Django>=2.2.5,<2.3.0
djangorestframework>=3.10.2,<3.11.0
flake8>=3.6.0,<3.7.0
pytest-django>=3.5.1,<3.6.0 
psycopg2-binary>=2.7.5,<2.8.0